Deck test refactoring
