+++ core/src/test/scala/it/parttimeteam/core/client/ClientInterfaceSpec.scala
@@ -1,0 +1,5 @@
+package it.parttimeteam.core.client
+
+class ClientInterfaceSpec {
+
+}
